requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
mysqlclient==2.2.0
tinydb==4.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
